# functions/api/schema/openapi.yaml
openapi: '3.1.0'
info:
  title: OpenPecha API v2
  version: '2.1.0'
servers:
  - url: https://api-l25bgmwqoa-uc.a.run.app
    description: Development
  - url: https://api-aq25662yyq-uc.a.run.app
    description: Production
paths:
...
  /v2/texts/{text_id}/instances:
    get:
      summary: Get instances for text
      description: Retrieve all instances associated with a text
      tags:
        - Instances
      parameters:
        - name: text_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the text
      responses:
        "200":
          description: Instances retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    type:
                      type: string
                      enum: [diplomatic, critical, collated]
                    copyright:
                      type: string
                      enum: [public, copyrighted]
                    annotations:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: string
                          type:
                            type: string
                            enum: [segmentation, alignment, version]
                          aligned_to:
                            type: string
                    bdrc:
                      type: string
                    wiki:
                      type: string
                    colophon:
                      type: string
                    incipit_title:
                      type: object
                      additionalProperties:
                        type: string
                    alt_incipit_titles:
                      type: array
                      items:
                        type: object
                        additionalProperties:
                          type: string
                    alignment_sources:
                      type: array
                      items:
                        type: string
                      description: List of manifestation IDs that have annotations aligned to this manifestation
                    alignment_targets:
                      type: array
                      items:
                        type: string
                      description: List of manifestation IDs that this manifestation's annotations are aligned to

              examples:
                success:
                  summary: List of instances
                  value:
                    - id: "I12345678"
                      type: "diplomatic"
                      copyright: "public"
                      annotations:
                        - id: "A12345678"
                          type: "segmentation"
                        - id: "A87654321"
                          type: "alignment"
                          aligned_to: "A11111111"
                      bdrc: "W123456"
                      alignment_sources: []
                      alignment_targets: ["T87654321", "T11111111"]
                      bdrc: "W123456"
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/ServerError'

    post:
      summary: Create a new instance
      description: Create a new instance with metadata and annotation for a specific text
      tags:
        - Instances
      parameters:
        - name: text_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the text to create an instance for
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - metadata
                - annotation
                - content
              properties:
                metadata:
                  type: object
                  required:
                    - type
                  properties:
                    type:
                      type: string
                      enum: [diplomatic, critical, collated]
                      description: Type of manifestation
                    copyright:
                      type: string
                      enum: [public, copyrighted]
                      default: public
                      description: Copyright status
                    bdrc:
                      type: string
                      description: BDRC identifier (required for diplomatic type)
                    wiki:
                      type: string
                      description: Wiki identifier
                    colophon:
                      type: string
                      description: Colophon text
                    incipit_title:
                      type: object
                      additionalProperties:
                        type: string
                      description: Localized incipit title
                    alt_incipit_titles:
                      type: array
                      items:
                        type: object
                        additionalProperties:
                          type: string
                      description: Alternative localized incipit titles
                annotation:
                  type: array
                  items:
                    type: object
                  description: Annotation data
                content:
                  type: string
                  description: The text content
            examples:
              valid_instance:
                summary: Valid instance creation request
                value:
                  metadata:
                    type: "diplomatic"
                    copyright: "public"
                    bdrc: "W123456"
                    colophon: "Sample colophon text"
                    incipit_title:
                      en: "Opening words"
                      bo: "དབུ་ཚིག"
                  annotation:
                    - span:
                        start: 0
                        end: 20
                      index: 0
                      alignment_index: [0]
                  content: "This is the text content to be stored"
      responses:
        "201":
          description: Instance created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      - "Instance created successfully"
                  id:
                    type: string
                    examples:
                      - "ABC12345678"
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "422":
          $ref: '#/components/responses/ValidationError'
        "500":
          $ref: '#/components/responses/ServerError'
...
  /v2/texts:
    get:
      summary: Retrieve all texts
      description: Fetch all texts with optional filtering and pagination
      tags:
        - Texts
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of results to skip
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [root, commentary, translation]
          description: Filter by text type
        - name: language
          in: query
          required: false
          schema:
            type: string
          description: Filter by language code
        - name: author
          in: query
          required: false
          schema:
            type: string
          description: Filter by author name
      responses:
        "200":
          description: Texts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    type:
                      type: string
                      enum: [root, commentary, translation]
                    title:
                      type: object
                      additionalProperties:
                        type: string
                    alt_titles:
                      type: array
                      items:
                        type: object
                        additionalProperties:
                          type: string
                    language:
                      type: string
                    contributions:
                      type: array
                      items:
                        type: object
                        properties:
                          person_id:
                            type: string
                          person_bdrc_id:
                            type: string
                          role:
                            type: string
                            enum: [author, translator, reviser]
                    date:
                      type: string
                    bdrc:
                      type: string
                    wiki:
                      type: string
                    parent:
                      type: string
              examples:
                success:
                  summary: List of texts
                  value:
                    - id: "ABC12345678"
                      type: "root"
                      title:
                        en: "Sample Expression"
                        bo: "དཔེ་མཚོན་ཚིག་སྒྲུབ།"
                      language: "bo"
                      contributions:
                        - person_id: "P12345678"
                          role: "author"
                      date: "2024-01-01"
                      bdrc: "W123456"
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "500":
          $ref: '#/components/responses/ServerError'

    post:
      summary: Create text
      description: Create a new text record
      tags:
        - Texts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - title
                - language
                - contributions
              properties:
                type:
                  type: string
                  enum: [root, commentary, translation]
                title:
                  type: object
                  additionalProperties:
                    type: string
                  description: Localized title
                alt_titles:
                  type: array
                  items:
                    type: object
                    additionalProperties:
                      type: string
                  description: Alternative localized titles
                language:
                  type: string
                  description: Primary language code
                contributions:
                  type: array
                  items:
                    type: object
                    properties:
                      person_id:
                        type: string
                      person_bdrc_id:
                        type: string
                      role:
                        type: string
                        enum: [author, translator, reviser]
                date:
                  type: string
                bdrc:
                  type: string
                wiki:
                  type: string
                parent:
                  type: string
                  description: Parent metadata ID (required for commentary/translation)
            examples:
              root_text:
                summary: Root text creation
                value:
                  type: "root"
                  title:
                    en: "Sample Root Text"
                    bo: "རྩ་བའི་གཞུང་དཔེ།"
                  language: "bo"
                  contributions:
                    - person_id: "P12345678"
                      role: "author"
                  date: "2024-01-01"
                  bdrc: "W123456"
              translation_text:
                summary: Translation text creation
                value:
                  type: "translation"
                  title:
                    en: "English Translation"
                  language: "en"
                  contributions:
                    - person_bdrc_id: "P87654321"
                      role: "translator"
                  parent: "T12345678"
      responses:
        "201":
          description: Text created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    examples:
                      - "Text created successfully"
                  id:
                    type: string
                    examples:
                      - "T12345678"
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "422":
          $ref: '#/components/responses/ValidationError'
        "500":
          $ref: '#/components/responses/ServerError'
  /v2/texts/{text_id}:
    get:
      summary: Retrieve text by ID
      description: Fetch a specific text by its ID
      tags:
        - Texts
      parameters:
        - name: text_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the text to retrieve
      responses:
        "200":
          description: Metadata retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  type:
                    type: string
                    enum: [root, commentary, translation]
                  title:
                    type: object
                    additionalProperties:
                      type: string
                  alt_titles:
                    type: array
                    items:
                      type: object
                      additionalProperties:
                        type: string
                  language:
                    type: string
                  contributions:
                    type: array
                    items:
                      type: object
                      properties:
                        person_id:
                          type: string
                        person_bdrc_id:
                          type: string
                        role:
                          type: string
                          enum: [author, translator, reviser]
                  date:
                    type: string
                  bdrc:
                    type: string
                  wiki:
                    type: string
                  parent:
                    type: string
              examples:
                success:
                  summary: Single text
                  value:
                    id: "T12345678"
                    type: "root"
                    title:
                      en: "Sample Expression"
                      bo: "དཔེ་མཚོན་ཚིག་སྒྲུབ།"
                    language: "bo"
                    contributions:
                      - person_id: "P12345678"
                        role: "author"
                    date: "2024-01-01"
                    bdrc: "W123456"
        "400":
          $ref: '#/components/responses/InvalidRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/ServerError'
...
components:
  responses:
    InvalidRequest:
      description: There was an error with the request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Error message
                examples:
                  - "There was an error with the request"
            required:
              - error
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Error message
                examples:
                  - "Resource was not found"
            required:
              - error
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        type: string
                    msg:
                      type: string
                    type:
                      type: string
            required:
              - error
              - details
          examples:
            fieldMissing:
              value:
                error: "Validation error"
                details: [
                  {
                    "loc": ["body", "title"],
                    "msg": "field required", 
                    "type": "value_error.missing"
                  }
                ]
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: Error message
                examples:
                  - "There was an error on the server"
            required:
              - error